test: expose executors
