--- tests/integration/capacity/flows/capacity.yml
@@ -1,12 +1,0 @@
-jtype: Flow
-executors:
-  - name: sentencizer0
-    uses: jinahub+docker://Sentencizer
-    jcloud:
-      capacity: spot
-  - name: sentencizer1
-    uses: jinahub+docker://Sentencizer
-    jcloud:
-      capacity: dedicated
-  - name: sentencizer2
-    uses: jinahub+docker://Sentencizer
