--- tests/integration/flows/resource-in-gateway.yml
@@ -1,13 +1,0 @@
-jtype: Flow
-with:
-  protocol: grpc
-jcloud:
-  gateway:
-    ingress: kong
-    resources:
-      requests:
-        memory: 800M
-        cpu: 0.4
-executors:
-  - name: sentencizer
-    uses: jinahub+docker://Sentencizer
